+++ lib/screens/bottom_nav_bar/profile/my_orders/show_rating_dialog_box.dart
@@ -47,7 +47,7 @@ class ShowRatingDialogBox extends StatelessWidget {
               orderId: orderId,
               productId: productId,
               comment: response.comment,
-              rating: response.rating.toInt());
+              rating: response.rating.toDouble());
         },
       );
     });
